requirements.txt
torch>=2.0.0
transformers==4.35.0
datasets==2.14.5
accelerate==0.24.1
python-dotenv
requests
gradio
gradio_client
